Fixes - check if base is empty object in createComponents - split tests to separate for Typography - rename createTypographyComponents to createTypoComponents
