+++ packages/typography/src/index.tsx
@@ -1,7 +1,4 @@
-export {
-  createTypographyComponents,
-  getStyleProps
-} from './createTypographyComponents';
+export { createTypoComponents, getStyleProps } from './createTypoComponents';
 export { TypographyBase } from './TypographyBase';
 export { fontWeight } from './fontWeight';
 export { TypographyBaseProps } from './typings';
